+++ addon/adapters/odata.js
@@ -1,15 +1,19 @@
 import { assert, debug } from '@ember/debug';
-import { isNone } from '@ember/utils';
+import { isNone, isEmpty } from '@ember/utils';
 import { get, computed } from '@ember/object';
 import { getOwner } from '@ember/application';
 import { Promise } from 'rsvp';
 import $ from 'jquery';
 import DS from 'ember-data';
+import { run } from '@ember/runloop';
+import { A } from '@ember/array';
 
 import SnapshotTransform from '../utils/snapshot-transform';
 import ODataQueryAdapter from '../query/odata-adapter';
 import { capitalize, camelize } from '../utils/string-functions';
 import { pluralize } from 'ember-inflector';
+import generateUniqueId from '../utils/generate-unique-id';
+import { getResponseMeta, getBatchResponses, parseBatchResponse } from '../utils/batch-queries';
 
 
 /**
@@ -24,6 +28,7 @@ import { pluralize } from 'ember-inflector';
 export default DS.RESTAdapter.extend({
   headers: computed(function() {
     return {
+      'OData-Version': '4.0',
       'Prefer': 'return=representation'
     };
   }),
@@ -48,6 +53,10 @@ export default DS.RESTAdapter.extend({
     @return {Promise}
   */
   query(store, type, query) {
+    if (!this.store) {
+      this.store = store;
+    }
+
     let url = this._buildURL(query.modelName);
     let builder = new ODataQueryAdapter(url, store);
     let data = builder.getODataQuery(query);
@@ -251,7 +260,7 @@ export default DS.RESTAdapter.extend({
       url = `${config.APP.backendUrls.api}`;
     }
 
-    const resultUrl =  `${url}/${actionName}`;
+    const resultUrl = `${url}/${actionName}`;
 
     return resultUrl;
   },
@@ -318,7 +327,8 @@ export default DS.RESTAdapter.extend({
   callEmberOdataAction(actionName, data, url, fields, store, modelName, successCallback, failCallback, alwaysCallback) {
     const resultUrl = this.generateActionUrl(actionName, data, url);
     return this._callAjax(
-      { data: JSON.stringify(data),
+      {
+        data: JSON.stringify(data),
         url: resultUrl,
         method: 'POST',
         contentType: 'application/json; charset=utf-8',
@@ -344,13 +354,182 @@ export default DS.RESTAdapter.extend({
   callAction(actionName, data, url, fields, successCallback, failCallback, alwaysCallback) {
     const resultUrl = this.generateActionUrl(actionName, data, url);
     return this._callAjax(
-      { data: JSON.stringify(data),
+      {
+        data: JSON.stringify(data),
         url: resultUrl,
         method: 'POST',
         contentType: 'application/json; charset=utf-8',
         dataType: 'json',
         xhrFields: isNone(fields) ? {} : fields
       }, null, null, successCallback, failCallback, alwaysCallback);
+  },
+
+  /**
+    A method to send batch update, create or delete models in single transaction.
+
+    It is recommended to create new models with identifiers, otherwise, after saving, the model object in the store will be created anew.
+
+    The array which fulfilled the promise may contain the following values:
+    - `new model object` - for records created without client id.
+    - `same model object` - for created, updated or unaltered records.
+    - `null` - for deleted records.
+
+    @method batchUpdate
+    @param {DS.Store} store The store.
+    @param {DS.Model[]|DS.Model} models Is array of models or single model for batch update.
+    @return {Promise} A promise that fulfilled with an array of models in the new state.
+  */
+  batchUpdate(store, models) {
+    if (isEmpty(models)) {
+      throw new Error('The models is empty.');
+    }
+
+    models = A(models);
+
+    const boundary = `batch_${generateUniqueId()}`;
+
+    let requestBody = '--' + boundary + '\r\n';
+
+    const changeSetBoundary = `changeset_${generateUniqueId()}`;
+    requestBody += 'Content-Type: multipart/mixed;boundary=' + changeSetBoundary + '\r\n\r\n';
+
+    let contentId = 0;
+    const modelsByContentId = {};
+    models.forEach((model) => {
+      let modelDirtyType = model.get('dirtyType');
+
+      if (!modelDirtyType) {
+        return;
+      }
+
+      requestBody += '--' + changeSetBoundary + '\r\n';
+      requestBody += 'Content-Type: application/http\r\n';
+      requestBody += 'Content-Transfer-Encoding: binary\r\n';
+
+      contentId++;
+      requestBody += 'Content-ID: ' + contentId + '\r\n\r\n';
+      modelsByContentId[contentId] = model;
+
+      const skipUnchangedAttrs = true;
+      const snapshot = model._createSnapshot();
+      SnapshotTransform.transformForSerialize(snapshot, skipUnchangedAttrs);
+
+      let modelHttpMethod = 'POST';
+      switch (modelDirtyType) {
+        case 'created':
+          modelHttpMethod = 'POST';
+          break;
+
+        case 'updated':
+          modelHttpMethod = skipUnchangedAttrs ? 'PATCH' : 'PUT';
+          break;
+
+        case 'deleted':
+          modelHttpMethod = 'DELETE';
+          break;
+
+        default:
+          throw new Error(`Unknown requestType: ${modelDirtyType}`);
+      }
+
+      if (!this.store) {
+        this.store = store;
+      }
+
+      const modelUrl =  this._buildURL(snapshot.type.modelName, modelDirtyType === 'created' ? undefined : model.get('id'));
+
+      requestBody += modelHttpMethod + ' ' + modelUrl + ' HTTP/1.1\r\n';
+      requestBody += 'Content-Type: application/json;type=entry\r\n';
+      requestBody += 'Prefer: return=representation\r\n\r\n';
+
+      // Don't need to send any data for deleting.
+      if (modelDirtyType !== 'deleted') {
+        const serializer = store.serializerFor(snapshot.type.modelName);
+        const data = {};
+        serializer.serializeIntoHash(data, snapshot.type, snapshot);
+        requestBody += JSON.stringify(data) + '\r\n';
+      }
+    });
+
+    requestBody += '--' + changeSetBoundary + '--\r\n';
+
+    requestBody += '\r\n--' + boundary + '--';
+
+    const url = `${this._buildURL()}/$batch`;
+
+    const headers = $.extend({}, this.get('headers'));
+    headers['Content-Type'] = `multipart/mixed;boundary=${boundary}`;
+    const httpMethod = 'POST';
+
+    const options = {
+      method: httpMethod,
+      headers,
+      dataType: 'text',
+      data: requestBody,
+    };
+
+    return new Promise((resolve, reject) => $.ajax(url, options).done((response, statusText, xhr) => {
+      const meta = getResponseMeta(xhr.getResponseHeader('Content-Type'));
+      if (meta.contentType !== 'multipart/mixed') {
+        return reject(new Error('Invalid response type.'));
+      }
+
+      const responses = getBatchResponses(response, meta.boundary);
+      if (responses.length !== 1) {
+        return reject(new Error('Invalid number of responses.'));
+      }
+
+      const { contentType, changesets } = parseBatchResponse(responses[0]);
+      if (contentType !== 'multipart/mixed') {
+        return reject(new Error('Invalid response type.'));
+      }
+
+      if (changesets.length === 1 && !this.isSuccess(changesets[0].meta.status)) {
+        return reject(new Error('Invalid response.'));
+      }
+
+      const result = [];
+      for (const model of models) {
+        const changeset = changesets.find(c => modelsByContentId[c.contentID] === model);
+        const modelClass = store.modelFor(model.constructor.modelName);
+        const serializer = store.serializerFor(modelClass.modelName);
+        switch (model.get('dirtyType')) {
+          case 'created':
+            if (changeset.meta.status !== 201) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            if (!model.get('id')) {
+              run(store, store.unloadRecord, model);
+            }
+
+            result.push(run(store, store.push, serializer.normalize(modelClass, changeset.body)));
+            break;
+
+          case 'updated':
+            if (changeset.meta.status !== 200) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            result.push(run(store, store.push, serializer.normalize(modelClass, changeset.body)));
+            break;
+
+          case 'deleted':
+            if (changeset.meta.status !== 204) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            run(store, store.unloadRecord, model);
+            result.push(null);
+            break;
+
+          default:
+            result.push(model);
+        }
+      }
+
+      resolve(result);
+    }).fail(reject));
   },
 
   /**
@@ -383,7 +562,7 @@ export default DS.RESTAdapter.extend({
           if (typeof successCallback.then === 'function') {
             if (!isNone(alwaysCallback)) {
               if (typeof alwaysCallback.then === 'function') {
-                successCallback(msg).then(() => {alwaysCallback(msg).then(resolve(msg));});
+                successCallback(msg).then(() => { alwaysCallback(msg).then(resolve(msg)); });
               } else {
                 successCallback(msg).then(alwaysCallback(msg)).then(resolve(msg));
               }
@@ -420,7 +599,7 @@ export default DS.RESTAdapter.extend({
           if (typeof failCallback.then === 'function') {
             if (!isNone(alwaysCallback)) {
               if (typeof alwaysCallback.then === 'function') {
-                failCallback(msg).then(() => {alwaysCallback(msg).then(reject(msg));});
+                failCallback(msg).then(() => { alwaysCallback(msg).then(reject(msg)); });
               } else {
                 failCallback(msg).then(alwaysCallback(msg)).then(reject(msg));
               }
@@ -508,6 +687,11 @@ export default DS.RESTAdapter.extend({
    */
   _appendIdToURL(id, url, modelName) {
     let encId = encodeURIComponent(id);
+
+    if (!this.store) {
+      throw new Error('No store.');
+    }
+
     let model = this.store.modelFor(modelName);
     if (model.idType === 'string') {
       encId = `'${encId}'`;
@@ -530,8 +714,12 @@ export default DS.RESTAdapter.extend({
   },
 
   deleteAllRecords(store, modelName, filter) {
+    if (!this.store) {
+      this.store = store;
+    }
+
     let url = this._buildURL(modelName);
-    let pathName  = this.pathForType(modelName);
+    let pathName = this.pathForType(modelName);
     let builder = new ODataQueryAdapter(url, store);
     let filterVelue = builder._buildODataFilters(filter);
     let filterQuery = !isNone(filterVelue) ? '$filter=' + filterVelue : '';
@@ -581,7 +769,7 @@ export default DS.RESTAdapter.extend({
       serializer.serializeIntoHash(data, type, snapshot);
     }
 
-    return this.ajax(url, httpMethod, { data: data }).then(function(response) {
+    return this.ajax(url, httpMethod, { data: data }).then(function (response) {
       return response;
     });
   }